requirements.txt
pkg-resources==0.0.0
PyJWT==1.7.1
asgiref==3.3.0
Django==3.0.5
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.2.4
